implemented offline image storage

+++ FlickrClient/ViewControllers/HomeVC.swift
@@ -23,6 +23,9 @@ class HomeVC: UICollectionViewController{
     var photoViewModels = [PhotoViewModel]()
     var parentArray = [[PhotoViewModel]]()
     var pages = [Page]()
+    var arrayOfPhotoEntities = [[PhotoEntity]]()
+    var imageArrayHasData = true
+    let reachable = "Network Connection Reachable"
     
     var pageNumber = 1
     var continuePagination = true
@@ -72,18 +75,14 @@ class HomeVC: UICollectionViewController{
     private func setupFetchRequest() {
         let fetchRequest: NSFetchRequest<Page> = Page.fetchRequest()
         fetchRequest.sortDescriptors = []
-        if let result = try? appDelegate.persistentContainer.viewContext.fetch(fetchRequest) {
-            self.pages = result
-            print("pages: ", pages.count)
-            for page in pages {
-                
-                let fetchRequest: NSFetchRequest<PhotoEntity> = PhotoEntity.fetchRequest()
-                let predicate = NSPredicate(format: "page == %@", page)
-                fetchRequest.sortDescriptors = []
-                fetchRequest.predicate = predicate
-                if let result = try? appDelegate.persistentContainer.viewContext.fetch(fetchRequest) {
-                    print("images: ", result.count)
-                }
+        if let result = try? appDelegate.persistentContainer.viewContext.fetch(fetchRequest) { self.pages = result }
+        for page in pages {
+            let fetchRequest: NSFetchRequest<PhotoEntity> = PhotoEntity.fetchRequest()
+            let predicate = NSPredicate(format: "page == %@", page)
+            fetchRequest.sortDescriptors = []
+            fetchRequest.predicate = predicate
+            if let images = try? appDelegate.persistentContainer.viewContext.fetch(fetchRequest) {
+                self.arrayOfPhotoEntities.append(images)
             }
         }
     }
@@ -92,12 +91,10 @@ class HomeVC: UICollectionViewController{
         NotificationCenter.default.addObserver(forName: NSNotification.Name("NetworkReachabilityChanged"), object: nil, queue: nil, using: {
             (notification) in
             if let userInfo = notification.userInfo {
-                if let messageTitle = userInfo["summary"] as? String, let reachableOrNot = userInfo["reachableOrNot"] as? String, let reachableStatus = userInfo["reachabilityStatus"] as? String {
-                    let messageFullBody = "\(reachableOrNot)\n\(reachableStatus)"
-                    let alertController = UIAlertController(title: messageTitle, message: messageFullBody, preferredStyle: .alert)
-                    let OKAction = UIAlertAction(title: "OK", style: .default)
-                    alertController.addAction(OKAction)
-                    self.present(alertController, animated: true, completion: nil)
+                if let reachableOrNot = userInfo["reachableOrNot"] as? String {
+                    if reachableOrNot == self.reachable {
+                        self.refresh()
+                    }
                 }
             }
         })
@@ -117,10 +114,12 @@ class HomeVC: UICollectionViewController{
         }
         
         if pageNumber == 3 { self.continuePagination = false }
+        if photos.count == 0 { imageArrayHasData = false }
+        else { imageArrayHasData = true }
         self.photoViewModels = photos.map({return PhotoViewModel(photo: $0)})
         parentArray.append(self.photoViewModels)
         
-        if pageNumber == 1 {
+        if pageNumber == 1 && imageArrayHasData {
             for page in pages {
                 appDelegate.persistentContainer.viewContext.delete(page)
                 try? appDelegate.persistentContainer.viewContext.save()
@@ -128,10 +127,12 @@ class HomeVC: UICollectionViewController{
             pages = []
         }
         
-        let page = Page(context: appDelegate.persistentContainer.viewContext)
-        page.pageNumber = Int64(pageNumber)
-        try? appDelegate.persistentContainer.viewContext.save()
-        pages.append(page)
+        if imageArrayHasData {
+            let page = Page(context: appDelegate.persistentContainer.viewContext)
+            page.pageNumber = Int64(pageNumber)
+            try? appDelegate.persistentContainer.viewContext.save()
+            pages.append(page)
+        }
         
         DispatchQueue.main.async {
             self.collectionView.reloadData()
@@ -140,13 +141,7 @@ class HomeVC: UICollectionViewController{
     }
 
     @objc func refreshTapped() {
-        photoViewModels = []
-        parentArray = []
-        setupActivityIndicator()
-        showActivityIndicator(show: true)
-        pageNumber = 1
-        downloadPhotos(page: pageNumber)
-        continuePagination = true
+        refresh()
     }
     
     private func loadingImages(_ loadingImages: Bool) {
@@ -173,6 +168,16 @@ class HomeVC: UICollectionViewController{
         }
     }
     
+    private func refresh() {
+        photoViewModels = []
+        parentArray = []
+        setupActivityIndicator()
+        showActivityIndicator(show: true)
+        pageNumber = 1
+        downloadPhotos(page: self.pageNumber)
+        continuePagination = true
+    }
+    
     // Pagination
     override func scrollViewDidEndDragging(_ scrollView: UIScrollView, willDecelerate decelerate: Bool) {
         let offsetY = scrollView.contentOffset.y
@@ -194,41 +199,62 @@ class HomeVC: UICollectionViewController{
 
 extension HomeVC {
     override func numberOfSections(in collectionView: UICollectionView) -> Int {
-        return parentArray.count
+        if imageArrayHasData { return parentArray.count }
+        else { return pages.count }
     }
     
     override func collectionView(_ collectionView: UICollectionView, numberOfItemsInSection section: Int) -> Int {
-        return parentArray[section].count
+        if imageArrayHasData { return parentArray[section].count }
+        else {
+            let page = pages[section]
+            let fetchRequest: NSFetchRequest<PhotoEntity> = PhotoEntity.fetchRequest()
+            let predicate = NSPredicate(format: "page == %@", page)
+            fetchRequest.sortDescriptors = []
+            fetchRequest.predicate = predicate
+            if let images = try? appDelegate.persistentContainer.viewContext.fetch(fetchRequest) {
+                return images.count
+            } else { return 0 }
+        }
     }
     
     override func collectionView(_ collectionView: UICollectionView, cellForItemAt indexPath: IndexPath) -> UICollectionViewCell {
         let cell = collectionView.dequeueReusableCell(withReuseIdentifier: cellId, for: indexPath) as! PhotoCell
         cell.imageView.image = UIImage(named: "placeholder")
-        
-        if indexPath.section < parentArray.count {
-            let photoViewModels = parentArray[indexPath.section]
-            
-            if indexPath.row < photoViewModels.count {
-                let photoViewModel = photoViewModels[indexPath.row]
-                FlickrClient.getImage(urlString: photoViewModel.url?.absoluteString ?? "") { (image, isImageFromCache) in
-                    if isImageFromCache == false {
-                        let page = self.pages[indexPath.section]
-                        let photoEntity = PhotoEntity(context: self.appDelegate.persistentContainer.viewContext)
-                        photoEntity.image = image?.pngData()
-                        photoEntity.urlString = photoViewModel.url?.absoluteString
-                        photoEntity.page = page
-                        try? self.appDelegate.persistentContainer.viewContext.save()
-                    }
-                    DispatchQueue.main.async {
-                        cell.imageView.image = image
+        guard imageArrayHasData == false else {
+            if indexPath.section < parentArray.count {
+                let photoViewModels = parentArray[indexPath.section]
+                
+                if indexPath.row < photoViewModels.count {
+                    let photoViewModel = photoViewModels[indexPath.row]
+                    FlickrClient.getImage(urlString: photoViewModel.url?.absoluteString ?? "") { (image, isImageFromCache) in
+                        if isImageFromCache == false {
+                            let page = self.pages[indexPath.section]
+                            let photoEntity = PhotoEntity(context: self.appDelegate.persistentContainer.viewContext)
+                            photoEntity.image = image?.pngData()
+                            photoEntity.urlString = photoViewModel.url?.absoluteString
+                            photoEntity.page = page
+                            try? self.appDelegate.persistentContainer.viewContext.save()
+                        }
+                        DispatchQueue.main.async {
+                            cell.imageView.image = image
+                        }
                     }
                 }
             }
+            return cell
         }
+        
+        let images = arrayOfPhotoEntities[indexPath.section]
+        if indexPath.row < images.count {
+            let image = images[indexPath.row]
+            cell.imageView.image = UIImage(data: image.image!)
+        }
+        
         return cell
     }
     
     override func collectionView(_ collectionView: UICollectionView, didSelectItemAt indexPath: IndexPath) {
+        guard imageArrayHasData else { return }
         let detailVC = DetailVC()
         let photoViewModels = parentArray[indexPath.section]
         detailVC.photoViewModel = photoViewModels[indexPath.row]
